requirements.txt
svgpathtools
pyserial
pyyaml
